+++ lib/chat_api_web/controllers/setup_status_controller.ex
@@ -1,4 +1,4 @@
-defmodule ChatApiWeb.GettingStartedStepsController do
+defmodule ChatApiWeb.SetupStatusController do
   use ChatApiWeb, :controller
 
   alias ChatApi.{Accounts, BrowserSessions, Github, Google, Mattermost, SlackAuthorizations, Twilio, Users}
